requirements.txt
# Discord Bot
discord.py>=2.3.2

# MongoDB (async)
motor>=3.3.2
pymongo>=4.6.3

# Data validation
pydantic>=2.6.0
pydantic-settings>=2.2.0

# Environment variables
python-dotenv>=1.0.1

# Logging / CLI
rich>=13.7.1

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.6
mongomock-motor>=0.0.21
